+++ release/flashlight/R/light_performance.R
@@ -12,9 +12,6 @@
 #' @param by An optional vector of column names used to additionally group the results. Will overwrite \code{x$by}.
 #' @param metrics An optional named list with metrics. Each metric takes at least four arguments: actual, predicted, case weights w and \code{...}.
 #' @param use_linkinv Should retransformation function be applied? Default is FALSE.
-#' @param metric_name Deprecated. Use \code{options(flashlight.metric_name = ...)}.
-#' @param value_name Deprecated. Use \code{options(flashlight.value_name = ...)}.
-#' @param label_name Deprecated. Use \code{options(flashlight.label_name = ...)}.
 #' @param ... Arguments passed from or to other functions.
 #' @return An object of class \code{light_performance}, \code{light} (and a list) with the following elements.
 #' \itemize{
@@ -40,23 +37,20 @@ light_performance.default <- function(x, ...) {
 
 #' @describeIn light_performance Model performance of flashlight object.
 #' @export
-light_performance.flashlight <- function(
-    x, data = x$data, by = x$by, metrics = x$metrics,
-    use_linkinv = FALSE, metric_name = NULL,
-    value_name = NULL, label_name = NULL, ...
-  ) {
+light_performance.flashlight <- function(x, data = x$data, by = x$by,
+                                         metrics = x$metrics,
+                                         use_linkinv = FALSE, ...) {
+  warning_on_names(c("metric_name", "value_name", "label_name"), ...)
 
-  warning_on_names(metric_name, value_name, label_name)
-
-  # Initialization
   metric_name <- getOption("flashlight.metric_name")
   value_name <- getOption("flashlight.value_name")
   label_name <- getOption("flashlight.label_name")
 
   stopifnot(
+    "No data!" = is.data.frame(data) && nrow(data) >= 1L,
+    "'by' not in 'data'!" = by %in% colnames(data),
     !anyDuplicated(c(metric_name, value_name, label_name, "pred_", by)),
     "No metric!" = !is.null(metrics),
-    "No data!" = is.data.frame(data) && nrow(data) >= 1L,
     "No 'y' defined in flashlight!" = !is.null(x$y)
   )