+++ release/flashlight/R/light_scatter.R
@@ -11,9 +11,6 @@
 #' @param use_linkinv Should retransformation function be applied? Default is TRUE. Not used for \code{type = "shap"}.
 #' @param n_max Maximum number of data rows to select. Will be randomly picked from the relevant data.
 #' @param seed An integer random seed used for subsampling.
-#' @param value_name Column name in resulting \code{data} containing the values belonging to \code{type}. Defaults to "value".
-#' @param label_name Column name in resulting \code{data} containing the label of the flashlight. Defaults to "label".
-#' @param type_name Column name in the resulting \code{data} with the plot \code{type}.
 #' @param ... Further arguments passed from or to other methods.
 #' @return An object of class \code{light_scatter}, \code{light} (and a list) with the following elements.
 #' \itemize{
@@ -21,9 +18,6 @@
 #'   \item \code{by} Same as input \code{by}.
 #'   \item \code{v} The variable evaluated.
 #'   \item \code{type} Same as input \code{type}. For information only.
-#'   \item \code{value_name} Same as input \code{value_name}.
-#'   \item \code{label_name} Same as input \code{label_name}.
-#'   \item \code{type_name} Same as input \code{type_name}.
 #' }
 #' @export
 #' @examples
@@ -48,31 +42,44 @@ light_scatter.default <- function(x, ...) {
 #' @describeIn light_scatter Variable profile for a flashlight.
 #' @export
 light_scatter.flashlight <- function(x, v, data = x$data, by = x$by,
-                                     type = c("predicted", "response", "residual", "shap"),
-                                     use_linkinv = TRUE, n_max = 400, seed = NULL,
-                                     value_name = "value", label_name = "label",
-                                     type_name = "type", ...) {
+                                     type = c("predicted", "response",
+                                              "residual", "shap"),
+                                     use_linkinv = TRUE, n_max = 400,
+                                     seed = NULL, ...) {
   type <- match.arg(type)
+
+  warning_on_names(c("value_name", "label_name"), ...)
+
+  value_name <- getOption("flashlight.value_name")
+  label_name <- getOption("flashlight.label_name")
 
   # If SHAP, extract data
   if (type == "shap") {
     if (!is.shap(x$shap)) {
       stop("No shap values calculated. Run 'add_shap' for the flashlight first.")
     }
-    data <- x$shap$data[x$shap$data[[x$shap$variable_name]] == v, ]
+    variable_name <- getOption("flashlight.variable_name")
+    data <- x$shap$data[x$shap$data[[variable_name]] == v, ]
   }
 
   # Checks
-  vars <- c(label_name, type_name, by, v, value_name)
-  stopifnot(v %in% colnames(data),
-            (n <- nrow(data)) >= 1,
-            !anyDuplicated(vars))
+  vars <- c(label_name, by, v, value_name)
+  stopifnot(
+    "No data!" = is.data.frame(data) && nrow(data) >= 1L,
+    "'by' not in 'data'!" = by %in% colnames(data),
+    "'v' not in 'data'!" = v %in% colnames(data),
+    !anyDuplicated(vars)
+  )
+  if (type %in% c("response", "residual") && !("y" %in% names(x))) {
+    stop("You need to specify 'y' in flashlight.")
+  }
+  n <- nrow(data)
 
   # Subsample rows if data too large
-  if (!is.null(seed)) {
-    set.seed(seed)
-  }
-  if (n_max < n) {
+  if (n > n_max) {
+    if (!is.null(seed)) {
+      set.seed(seed)
+    }
     data <- data[sample(n, n_max), , drop = FALSE]
   }
 
@@ -92,13 +99,10 @@ light_scatter.flashlight <- function(x, v, data = x$data, by = x$by,
 
   # Organize output
   data[[label_name]] <- x$label
-  data[[type_name]] <- type
-
-  out <- list(data = as_tibble(data[, vars]), by = by, v = v,
-              type = type, value_name = value_name,
-              label_name = label_name, type_name = type_name)
-  class(out) <- c("light_scatter", "light", "list")
-  out
+  add_classes(
+    list(data = as_tibble(data[, vars]), by = by, v = v, type = type),
+    c("light_scatter", "light")
+  )
 }
 
 #' @describeIn light_scatter light_scatter for a multiflashlight.