+++ release/flashlight/R/most_important.R
@@ -28,9 +28,15 @@ most_important.default <- function(x, top_m = Inf) {
 #' @describeIn most_important Extracts most important variables from an object of class \code{light_importance}.
 #' @export
 most_important.light_importance <- function(x, top_m = Inf) {
-  data <- group_by(x$data, across(x$variable_name))
-  total_importance <- summarize(data, across(x$value_name, sum, na.rm = TRUE),
-                                .groups = "drop")
-  total_importance <- arrange(total_importance, across(x$value_name, desc))
-  head(total_importance[[x$variable_name]], top_m)
+  value_name <- getOption("flashlight.value_name")
+  variable_name <- getOption("flashlight.variable_name")
+
+  data <- group_by(x$data, across(variable_name))
+  total_importance <- summarize(
+    data, across(all_of(value_name), sum, na.rm = TRUE),
+    .groups = "drop"
+  )
+  total_importance <- arrange(total_importance,
+                              across(all_of(value_name), desc))
+  head(total_importance[[variable_name]], top_m)
 }